fix: memory management

+++ 0x15-file_io/0-read_textfile.c
@@ -15,8 +15,8 @@ ssize_t read_textfile(const char *filename, size_t letters)
 {
 	char *buf;
 	size_t ans;
-	int test_read;
-	int fd = open(filename, O_RDONLY, 00777);
+	ssize_t test_read;
+	ssize_t fd = open(filename, O_RDONLY, 00777);
 
 	buf = malloc(letters);
